Events > EventCard > Card component flow implemented
