+++ client/src/components/Events.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import SwipeableViews from 'react-swipeable-views';
-import EventDetail from './EventDetail';
+import EventCards from './EventCards';
+import EventGrid from './EventGrid';
 import { returnToSearch } from '../actions';
 
 class Events extends Component {
@@ -17,17 +17,6 @@ class Events extends Component {
   }
 
   render() {
-
-    const showEvents = this.state.eventsToDisplay.map((event, index) => {
-      return <EventDetail 
-                listView={this.state.listView}
-                style={Object.assign({}, slideStyles.slide)}
-                key={event.id}
-                event={event}
-                active={index + 1}
-                length={this.state.eventsToDisplay.length}
-              />;
-    })
 
     return (
       <div>
@@ -45,21 +34,17 @@ class Events extends Component {
             </div>
           </div>
           { this.state.eventsToDisplay.length <= 0 ?
-            <p>Sorry, no events.</p> : null
+            <p className="noEventsFound">Sorry, no events.</p> : null
           }
-          <button className="previousEventButton" onClick={this.previousEvent}>&lt;</button>
-          <button className="nextEventButton" onClick={this.nextEvent}>&gt;</button>
           { this.state.listView === false ?
-            <SwipeableViews className="EventDetail" index={this.state.activeEventIndex} onChangeIndex={this.handleChangeIndex}>
-              {showEvents}
-            </SwipeableViews>
+            <EventCards events={this.state.eventsToDisplay}/>
           :
-            <div className="EventDetailGrid">{showEvents}</div>
+            <EventGrid events={this.state.eventsToDisplay}/>
           }
         </div>
       </div>
     );
-  }
+  } 
 
   listToggleHandle = (event) => {
     this.setState({
@@ -111,44 +96,12 @@ class Events extends Component {
     })
   }
 
-  nextEvent = (event) => {
-    event.preventDefault();
-    if (this.state.activeEventIndex === this.state.eventsToDisplay.length - 1) {
-      return;
-    } else {
-      let newIndex = this.state.activeEventIndex + 1;
-      this.setState({
-        activeEventIndex: newIndex
-      })
-    }
-  }
-
-  previousEvent = (event) => {
-    event.preventDefault();
-    if (this.state.activeEventIndex === 0) {
-      return;
-    } else {
-      let newIndex = this.state.activeEventIndex - 1;
-      this.setState({
-        activeEventIndex: newIndex
-      })
-    }
-  }
-
   handleChangeIndex = (index) => {
     this.setState({
       activeEventIndex: index
     })
   }
 
-}
-
-const slideStyles = {
-  slide: {
-    padding: 15,
-    minHeight: 100,
-    color: "#fff"
-  }
 }
 
 function sanitizeString(string) {
